% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda_n_gene}
\alias{eda_n_gene}
\title{Plot the expression of one or two genes with an optional stratifier}
\usage{
eda_n_gene(dds, genes, stratifier = NULL, assay = 2)
}
\arguments{
\item{dds}{A \code{DESeqDataSet} object}

\item{genes}{A character vector of one or two genes.}

\item{stratifier}{A name of a column of the \code{dds} \code{colData} that will be used
to distinguish groups - see the dispatched function documentation for more
details (see details)}

\item{assay}{integer. Assay position in the supplied \code{dds} that corresponds
to normalized counts.}
}
\value{
a ggplot
}
\description{
Plot the expression of one or two genes with an optional stratifier
}
\details{
\code{eda_n_gene} will dispatch either \code{eda_one_gene} or \code{eda_two_gene}
depending on the length of the \code{genes} argument.
}
