% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda_two_gene}
\alias{eda_two_gene}
\title{Plot the expression of two genes}
\usage{
eda_two_gene(dds, gene_x, gene_y, stratifier = NULL, assay = 2)
}
\arguments{
\item{dds}{A \code{DESeqDataSet} object}

\item{gene_x}{character. An HGNC symbol, like "KRT14", to be plotted on the
x-axis}

\item{gene_y}{character. An HGNC symbol, like "SRC", to be plotted on the
y-axis}

\item{stratifier}{A name of a column of the \code{dds} \code{colData} that will be used
to color the points of the scatterplot}

\item{assay}{integer. Assay position in the supplied \code{dds} that corresponds
to normalized counts.}
}
\value{
a ggplot
}
\description{
Plot the expression of two genes
}
