% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{get_gene_expression}
\alias{get_gene_expression}
\title{Get expression of gene by indices in a dds}
\usage{
get_gene_expression(dds, gene_indices, assay)
}
\arguments{
\item{dds}{A DESeqDataSet}

\item{gene_indices}{integer vector that refer to the row number of genes to retrieve}

\item{assay}{integer, assay index that refers to the desired expression data slot}
}
\value{
A dataframe where row names are sample names, and column names are
gene names
}
\description{
Get expression of gene by indices in a dds
}
\details{
Importantly, returns data as a \code{data.frame} without dropping
dimensions.
}
